modified lexica for amt

+++ FFN_SVM/FFN/generation/deep.py
@@ -1,5 +1,7 @@
 import logging
 from typing import *
+import os
+
 import pandas as pd
 
 import torch
@@ -27,9 +29,13 @@ def deep(
     NNnet.load_state_dict(torch.load(args.model_dir+"/"+train+".bin"))
     trainDf, devDf, _ = splitData(getData(args.dataFolder, train))
     
-    lexicon = generateLexicon_FFN(NNnet,trainDf,nlp,args.method,background=background,device=device)
-    outfilename = f"{args.output_dir}/{train}_ffn_deep.csv"
-    lexicon.to_csv(outfilename, index = False, index_label = False)
+    outfilename = f"{args.output_dir}/{train}_ffn_ds.csv"
+    if os.path.exists(outfilename):
+        logger.info("File already exists, skipped!")
+        lexicon = pd.read_csv(outfilename)
+    else:
+        lexicon = generateLexicon_FFN(NNnet,trainDf,nlp,args.method,background=background,device=device)
+        lexicon.to_csv(outfilename, index = False, index_label = False)
     lexiconWords, lexiconMap = getLexicon(df = lexicon)
     
     logger.info("Running evaluation.")
